requirements.txt
google-generativeai==0.8.5
numpy==1.26.4
faiss-cpu==1.11.0.post1
psycopg2-binary==2.9.9
flask==2.3.3
pandas==2.1.3
python-dotenv==1.0.0